setting auth token in cookies
