+++ client/app/dashboard/page.tsx
@@ -1,0 +1,11 @@
+import { Button } from "@/components/ui/button";
+import Image from "next/image";
+import Link from "next/link";
+
+export default function Home() {
+  return (
+    <div className="p-8 flex justify-between">
+       
+    </div>
+  );
+}
