--- client/app/login/page.tsx
@@ -1,83 +1,0 @@
-"use client"
-
-import { zodResolver } from "@hookform/resolvers/zod"
-import * as z from "zod"
-import { useForm } from "react-hook-form"
-import { Button } from "@/components/ui/button"
-import {
-    Form,
-    FormControl,
-    FormDescription,
-    FormField,
-    FormItem,
-    FormLabel,
-    FormMessage,
-} from "@/components/ui/form"
-import { Input } from "@/components/ui/input"
-import getUserToken from "@/actions/userToken"
-import { redirect } from "next/navigation"
-import { useRouter } from "next/navigation"
-
-const formSchema = z.object({ "name": z.string().min(1).max(255), "password": z.string().min(1).max(255) })
-
-export default function LoginForm() {
-
-    const router = useRouter()
-
-    const form = useForm<z.infer<typeof formSchema>>({
-        resolver: zodResolver(formSchema),
-        defaultValues: {
-            name: "",
-            password: "",
-        },
-    })
-
-    async function onSubmit(values: z.infer<typeof formSchema>) {
-        console.log(values);
-        
-        const userToken = await getUserToken(values.name, values.password)
-        router.push("/")
-    }
-
-    return (
-        <div className="flex p-16">
-            <Form {...form}>
-                <form noValidate onSubmit={form.handleSubmit(onSubmit)} className="space-y-8">
-
-                    <FormField
-                        control={form.control}
-                        name="name"
-                        render={({ field }) => (
-                            <FormItem>
-                                <FormLabel>UserName</FormLabel>
-                                <FormControl>
-                                    <Input placeholder="userName " {...field} />
-                                </FormControl>
-                                <FormDescription>
-                                    Taiga userName
-                                </FormDescription>
-                            </FormItem>
-                        )}
-                    />
-
-                    <FormField
-                        control={form.control}
-                        name="password"
-                        render={({ field }) => (
-                            <FormItem>
-                                <FormLabel>Password</FormLabel>
-                                <FormControl>
-                                    <Input type="password" placeholder="Password" {...field} />
-                                </FormControl>
-                                <FormDescription>
-                                    Taiga password
-                                </FormDescription>
-                            </FormItem>
-                        )}
-                    />
-                    <Button type="submit">Submit</Button>
-                </form>
-            </Form>
-        </div>
-    )
-}
